Changed width of images, change scroll button text and project link dropdown button.
Changed width of images to fit mobile better.
Scroll button arrow is now different, changed to be just a > rotated and larger.
To ProkectLink, made only the header a button, not entire element. Thus only can show and hide using header. Added a border to indicate that it is clickable.